All done
Created article page and added more content to the about us page. In the home page recipes were added to get more content and a buttons got the functionality to navigate to the right pages. 
In the index.css file, the navbar got editted a little bit so it looks better :)